Accelerate, refactor, fix
Add a configuration parameter for bypassing standard loading and
delegating loading of application class to internal loader to accelerate
execution. Refactored Algo_INIT and ExecutionContext to better isolate
members in ExecutionContext that must be accessed only during creation
of initial state (and removed a couple of incorrect uses). Fixed
finalization of State so that it does not close the stdin/out/err files.