Fix bug
Array classes shall not override Object.clone() method, otherwise
invokevirtual [Ljava/lang/Object:()Ljava/lang/Object;:clone does not
work when the 'this' object is, e.g., [Ljava/lang/Thread (that is a
subclass of java/lang/Object but not of [Ljava/lang/Object, although
assignment-compatible with the latter). Removed therefore all the
overriding of the Object.clone() method in array classes, and managed
the fact that the clone method in arrays is public.