Fix breadth mode for pre-initial execution
The use of symbolic KlassPseudoReferences as the output of some
sun.misc.Unsafe methods makes potentially spill these symbolic values on
the operand stack during the pre-initial phase as the <clinit> of the
target class is executed. With some breadth modes this creates branch
points during the pre-initial phase. To avoid it this fix sets to
MORE_THAN_ONE the breadth mode during the pre-initial phase.