Allow using Signature to set a method signature
Given the wide usage of Signature in the codebase, it might be useful to
allow setting up the symbolic method to execute relying directly on a
Signature.
The use case is when you are dealing with a large number of methods to
execute symbolically, and you have already constructed Signatures for
other purposes. This small patch allows to reuse already-constructed
objects.

Signed-off-by: Alessandro Pellegrini <a.pellegrini@ing.uniroma2.it>