Fix bug
The DecisionProcedureGuidanceJDI.getJDIValueField method assumed that,
when getting a field value in the concrete state, the field is declared
in the same class as the corresponding field in the symbolic state. This
can be invalidated by model classes, and is especially true for the
model class of the LinkedHashMap: While in the (real) LinkedHashMap the
size field used is the one declared in the HashMap superclass, the model
class JAVA_LINKEDMAP must redeclare a size field in the subclass to be
able to freely manipulate its value. With this commit we deleted the
check in the getJDIValueField method when looking for the field in the
concrete state on the class where this field shall be found.

In general, it is not a good idea to have symbolic field values of a
model class in the path condition, since there might be no guarantee
that the real class will have that field. A better idea could be using
symbolic function applications also for size (i.e., using size() as a
symbol for the map's size), as is already done for containment.